infra: move tests inside workspaces and add unit to CI
