--- .github/workflows/server-function.yml
@@ -1,49 +1,0 @@
-name: ci
-
-on:
-  pull_request:
-    branches: [main]
-  push:
-    branches: [main]
-
-jobs:
-  e2e-tests:
-    name: "E2E tests using ${{ matrix.browser }}"
-    strategy:
-      fail-fast: false
-      matrix:
-        browser: [chrome, firefox]
-
-    runs-on: ubuntu-22.04
-    steps:
-      - name: Checkout
-        uses: actions/checkout@v4
-      - uses: actions/setup-node@v4
-        with:
-          node-version-file: ".nvmrc"
-
-      - uses: pnpm/action-setup@v4
-        with:
-          # https://github.com/pnpm/pnpm/issues/8953
-          version: 9.15.3
-
-      - name: Install project dependencies
-        run: pnpm i
-
-      - name: Install project dependencies
-        run: pnpm run build
-
-      - name: Install dependencies
-        uses: cypress-io/github-action@v6
-        with:
-          working-directory: .
-          runTests: false
-
-      - name: Cypress test
-        uses: cypress-io/github-action@v6
-        with:
-          install: false
-          working-directory: tests/server-function
-          build: pnpm run build
-          start: pnpm run start
-          browser: ${{ matrix.browser }}
