+++ .github/workflows/tests.yml
@@ -1,0 +1,53 @@
+name: 🧪 Tests
+
+on:
+  workflow_dispatch:
+  pull_request:
+    branches: [main]
+  push:
+    branches: [main]
+
+jobs:
+  tests:
+    name: "🧪 Tests"
+    runs-on: ubuntu-22.04
+    steps:
+      - uses: actions/checkout@v3
+      - uses: pnpm/action-setup@v4
+        with:
+          # https://github.com/pnpm/pnpm/issues/8953
+          version: 9.15.3
+
+      - name: Install project dependencies
+        run: pnpm i
+
+      - name: Build SolidStart
+        run: pnpm --filter @solidjs/start build
+
+      - name: Build Test Project
+        run: pnpm --filter tests build
+
+      - name: Run unit tests
+        run: pnpm --filter tests unit:ci
+
+      - name: Install Cypress dependencies
+        uses: cypress-io/github-action@v6
+        with:
+          working-directory: .
+          runTests: false
+
+      - name: E2E Firefox
+        uses: cypress-io/github-action@v6
+        with:
+          install: false
+          working-directory: packages/tests
+          start: pnpm start
+          browser: firefox
+
+      - name: E2E Chromium
+        uses: cypress-io/github-action@v6
+        with:
+          install: false
+          working-directory: packages/tests
+          start: pnpm start
+          browser: chromium
