+++ packages/tests/cypress/screenshots/server-function.cy.ts/server-function -- should externalize node builtin in server function - toplevel (failed) (attempt 2).png
