+++ packages/tests/cypress/screenshots/server-function.cy.ts/server-function -- should have isServer true in the server function - toplevel (failed) (attempt 2).png
