+++ packages/tests/public/favicon.ico
