+++ packages/tests/src/routes/server-function-meta-nested.tsx
