+++ packages/tests/src/routes/treeshaking/(no-side-effects).tsx
