+++ packages/tests/src/routes/treeshaking/side-effects.tsx
