+++ packages/tests/vitest.config.ts
@@ -1,0 +1,10 @@
+import solid from "vite-plugin-solid";
+import { defineConfig } from "vitest/config";
+
+export default defineConfig({
+  plugins: [solid()],
+  test: {
+    mockReset: true,
+    environment: "node"
+  }
+});
