--- tests/server-function/vitest.config.ts
@@ -1,21 +1,0 @@
-import solid from "vite-plugin-solid";
-import { defineConfig } from "vitest/config";
-
-export default defineConfig({
-  plugins: [solid()],
-  test: {
-    mockReset: true,
-    environment: "node"
-
-    // at least one instance is required
-    // browser: {
-    //   provider: "playwright",
-    //   enabled: true,
-    //   instances: [
-    //     {
-    //       browser: "chromium"
-    //     }
-    //   ]
-    // }
-  }
-});
